wrok on database implementation

+++ logging/LDatabase.h
@@ -42,8 +42,8 @@ namespace LOGMASTER
 	{
 	public:
 		std::uint32_t m_id = 0;
-		std::uint8_t  m_level = 0;
-		std::uint8_t  m_category = 0;
+		std::uint32_t  m_level = 0;
+		std::uint32_t  m_category = 0;
 		double        m_time = 0;
 		std::string   m_json = "";
 	};
